Channels refactoring
