+++ ChitChat02/PresentationLayer/Modules/ConversationList/ChannelsViewController.xib
@@ -11,7 +11,6 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ChannelsViewController" customModule="ChitChat02" customModuleProvider="target">
             <connections>
                 <outlet property="channelsTableView" destination="KPf-eo-hwB" id="lww-ff-3u7"/>
-                <outlet property="emptyLabel" destination="9AX-Lp-O6M" id="Lof-55-L8u"/>
                 <outlet property="loadingIndicator" destination="nzt-MW-EBP" id="CgH-3V-Rcj"/>
                 <outlet property="view" destination="iN0-l3-epB" id="TmG-kT-QeQ"/>
             </connections>
@@ -28,25 +27,13 @@
                 <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" style="gray" translatesAutoresizingMaskIntoConstraints="NO" id="nzt-MW-EBP">
                     <rect key="frame" x="197" y="438" width="20" height="20"/>
                 </activityIndicatorView>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="9AX-Lp-O6M">
-                    <rect key="frame" x="186" y="438" width="42" height="20.5"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
             </subviews>
             <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
             <constraints>
                 <constraint firstItem="KPf-eo-hwB" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="7Mn-PF-Jhn"/>
-                <constraint firstItem="9AX-Lp-O6M" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="Rq1-Bg-IHj"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="9AX-Lp-O6M" secondAttribute="bottom" constant="16" id="Sx7-3h-xZR"/>
-                <constraint firstItem="9AX-Lp-O6M" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="VUq-i5-4ip"/>
-                <constraint firstItem="9AX-Lp-O6M" firstAttribute="top" relation="greaterThanOrEqual" secondItem="vUN-kp-3ea" secondAttribute="top" constant="16" id="Xdy-H2-SRh"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="9AX-Lp-O6M" secondAttribute="trailing" constant="16" id="ZXE-x9-iHH"/>
                 <constraint firstItem="KPf-eo-hwB" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="c2o-My-IiA"/>
                 <constraint firstItem="nzt-MW-EBP" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="lGy-Nt-kRf"/>
                 <constraint firstItem="nzt-MW-EBP" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="orC-8d-4kl"/>
-                <constraint firstItem="9AX-Lp-O6M" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="uR2-SQ-wdf"/>
                 <constraint firstItem="KPf-eo-hwB" firstAttribute="trailing" secondItem="vUN-kp-3ea" secondAttribute="trailing" id="xGS-Xj-XT9"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="KPf-eo-hwB" secondAttribute="bottom" id="yuK-dZ-5bn"/>
             </constraints>
